📦 NEW: refactor: remove unused types and improve location handling
- Deleted unused type definitions from types.ts.
- Updated add.vue to reset submit error on form submission.
- Changed migration output directory in drizzle.config.ts.
- Added devProxy configuration in nuxt.config.ts for local development.
- Refactored location API to use new query functions for creating and finding unique slugs.
- Added initial migration files for account, session, user, verification, location, locationLog, and locationLogImage tables.
- Created unique index for userId and name in location table.
- Updated migration metadata to reflect the new schema changes.
- Implemented a new query file for location operations, including slug generation and location creation.
- Reintroduced type definitions in types.d.ts for H3EventContext.
