+++ drizzle.config.ts
@@ -4,7 +4,7 @@ import { defineConfig } from "drizzle-kit";
 import env from "./app/lib/env";
 
 export default defineConfig({
-  out: "./app/lib/migrations",
+  out: "./app/lib/db/migrations",
   schema: "./app/lib/db/schema/index.ts",
   dialect: "turso",
   dbCredentials: {
