+++ server/api/locations/index.post.ts
@@ -1,11 +1,5 @@
-import { eq } from "drizzle-orm";
-import { customAlphabet } from "nanoid/non-secure";
-import slugify from "slug";
-
-import db from "~/lib/db";
-import { InsertLocation, location } from "~/lib/db/schema";
-
-const nanoid = customAlphabet("1234567890abcdef", 5);
+import { createLocation, findUniqueSlugLocation } from "~/lib/db/queries/location";
+import { InsertLocation } from "~/lib/db/schema";
 
 export default defineEventHandler(async (event) => {
   if (!event.context.user) {
@@ -33,32 +27,10 @@ export default defineEventHandler(async (event) => {
       data,
     }));
   }
-
-  let slug = slugify(result.data.name);
-  let existsLocation = await db.query.location.findFirst({
-    where: eq(location.slug, slug),
-  });
-
-  while (existsLocation) {
-    const id = nanoid();
-    const uniqueSlug = `${slug}-${id}`;
-
-    existsLocation = await db.query.location.findFirst({
-      where: eq(location.slug, uniqueSlug),
-    });
-
-    if (!existsLocation) {
-      slug = uniqueSlug;
-    }
-  }
+  const slug = await findUniqueSlugLocation(result.data.name);
 
   try {
-    const created = await db.insert(location).values({
-      ...result.data,
-      slug,
-      userId: event.context.user.id,
-    }).returning().get();
-
+    const created = await createLocation(result, slug, event.context.user.id);
     return created;
   }
   catch {
